Use portable format specifiers: Case %ld - plain - fprintf.
Fix uses of plain "%ld" within fprintf():
- Replace "%ld" with "%" PRId64.
- Cast corresponding argument to (int64_t).

+++ src/option.c
@@ -6447,7 +6447,7 @@ static int put_setnum(FILE *fd, char *cmd, char *name, long *valuep)
     /* print 'wildchar' and 'wildcharm' as a key name */
     if (fputs((char *)get_special_key_name((int)wc, 0), fd) < 0)
       return FAIL;
-  } else if (fprintf(fd, "%ld", *valuep) < 0)
+  } else if (fprintf(fd, "%" PRId64, (int64_t)*valuep) < 0)
     return FAIL;
   if (put_eol(fd) < 0)
     return FAIL;
